Renamed / Aligned package names
