+++ connection-generator/src/main/java/edu/kit/kastel/mcse/ardoco/core/connectiongenerator/extractors/InitialConnectionAgent.java
@@ -1,4 +1,4 @@
-package edu.kit.kastel.mcse.ardoco.core.connectiongenerator.agents_extractors;
+package edu.kit.kastel.mcse.ardoco.core.connectiongenerator.extractors;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,6 +6,7 @@ import java.util.Map;
 
 import org.kohsuke.MetaInfServices;
 
+import edu.kit.kastel.mcse.ardoco.core.connectiongenerator.GenericConnectionConfig;
 import edu.kit.kastel.mcse.ardoco.core.datastructures.agents.Configuration;
 import edu.kit.kastel.mcse.ardoco.core.datastructures.agents.ConnectionAgent;
 import edu.kit.kastel.mcse.ardoco.core.datastructures.agents.Loader;
