+++ recommendation-generator/src/main/java/edu/kit/kastel/mcse/ardoco/core/recommendationgenerator/extractors/NameTypeExtractor.java
@@ -1,4 +1,4 @@
-package edu.kit.kastel.mcse.ardoco.core.recommendationgenerator.agents_extractors;
+package edu.kit.kastel.mcse.ardoco.core.recommendationgenerator.extractors;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -17,6 +17,7 @@ import edu.kit.kastel.mcse.ardoco.core.datastructures.definitions.IRecommendatio
 import edu.kit.kastel.mcse.ardoco.core.datastructures.definitions.ITextState;
 import edu.kit.kastel.mcse.ardoco.core.datastructures.definitions.IWord;
 import edu.kit.kastel.mcse.ardoco.core.datastructures.extractors.RecommendationExtractor;
+import edu.kit.kastel.mcse.ardoco.core.recommendationgenerator.GenericRecommendationConfig;
 
 /**
  * This analyzer searches for name type patterns. If these patterns occur recommendations are created.
