+++ text-extractor/src/main/java/edu/kit/kastel/mcse/ardoco/core/textextractor/GenericTextConfig.java
@@ -1,4 +1,4 @@
-package edu.kit.kastel.mcse.ardoco.core.textextractor.agents_extractors;
+package edu.kit.kastel.mcse.ardoco.core.textextractor;
 
 import java.util.List;
 import java.util.Map;
