+++ text-extractor/src/main/java/edu/kit/kastel/mcse/ardoco/core/textextractor/extractors/InitialTextAgent.java
@@ -1,4 +1,4 @@
-package edu.kit.kastel.mcse.ardoco.core.textextractor.agents_extractors;
+package edu.kit.kastel.mcse.ardoco.core.textextractor.extractors;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -14,6 +14,7 @@ import edu.kit.kastel.mcse.ardoco.core.datastructures.definitions.ITextState;
 import edu.kit.kastel.mcse.ardoco.core.datastructures.definitions.IWord;
 import edu.kit.kastel.mcse.ardoco.core.datastructures.extractors.IExtractor;
 import edu.kit.kastel.mcse.ardoco.core.datastructures.extractors.TextExtractor;
+import edu.kit.kastel.mcse.ardoco.core.textextractor.GenericTextConfig;
 
 /**
  * The Class InitialTextAgent defines the agent that executes the extractors for the text stage.
